Used case_number in db
There is a case_number variable that holds on to the case number instead of selecting string elements.  This is a much more robust method for getting an HP number in the correct ##-##### format.

I question why you have functions to redefine functions (i.e meta_to_db) ; it seems silly and adds extra code.